backend/pyproject.toml
[tool.poetry]
name = "literacyai-dev"
version = "0.1.0"
description = ""
authors = ["Yshgodd <kyun20087@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
langchain = "^0.1.16"
langchain-community = "^0.0.34"
langchain-openai = "^0.1.3"
langchain-core = "^0.1.45"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
bs4 = "^0.0.2"
sqlalchemy = "^2.0.29"
psycopg2 = "^2.9.9"
passlib = "^1.7.4"
bcrypt = "^4.1.2"
apscheduler = "^3.10.4"
beautifulsoup4 = "^4.12.3"
requests = "^2.31.0"
black = "^24.4.2"
pytest = "^8.2.1"
pyjwt = "^2.8.0"
python-multipart = "^0.0.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
